requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
passlib[bcrypt]
python-jose
python-dotenv
